Erlang Runtime System: make compatible with unikernel environment
The first patch modifies BEAM so that it incorporates the
functionalities of erlexec and spawns the process mapper daemon as
a thread.
The second patch is needed to prevent a "Cookie file
/.erlang.cookie must be accessible by owner only" error that is
triggered when erlexec is called wit the "-name" or "-sname"
command line options (e.g. "-sname name@host").