Thread-based erl_child_setup alternative
erlang/OTP uses an external program called "erl_child_setup" to handle the
spawning of (and communication with) external processes. One of the principal
uses of this program is to in turn execute the "inet_gethost" program, which
is used to resolve internet addresses. To enable support for erlang/OTP under
Nanos, these two programs are internalized here as pthread-based versions.
Currently, only inet_gethost is emulated, but command lines for other programs
may be parsed and handled by thread-based versions of those programs if
necessary. These thread-based alternatives will only be used if the host
operating system name (via uname(2)) is set to "Nanos", and otherwise the
behavior will remain unchanged.