erlang patches for miner, fix writes in erl_child_setup_thread() (#5)
* erl_child_setup_thread: use write_exact()

erl_child_setup_thread() was calling write() for pipes without handling the
partial write case, EINTR, etc. This change borrows write_exact() from
erl_child_setup.c and uses that in place of calls to write().

* additional erlang patches to run helium miner

This adds two new patches to erlang/OTP: one to allow the modification of the
node name to use the configured address of a network interface, and another to
interpret and emulate shell commands used by the "inet_ext" package to find
the default gateway address.